small fixes, implemented query testing in histogram inspector

+++ TwigQuerySelectivity23/src/hist/Equidepth.java
@@ -27,7 +27,7 @@ public class Equidepth
             this.bars[i] = new Bar();
     }
     
-    public double getBarCount()
+    public int getBarCount()
     {
         return this.barCount;
     }
